change statTile and statBlock to ts
