+++ client/src/components/profile/deckEditor/StatBlock.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+interface props {
+  cardProperty: string;
+  cardStats: string;
+}
+
+const StatBlock: React.FC<props> = ({ cardProperty, cardStats }) => {
+  return (
+    <div className="stat-block">
+      {cardProperty} : {cardStats}%
+    </div>
+  );
+};
+
+export default StatBlock;
