+++ client/typings/custom/stat.d.ts
@@ -1,0 +1,4 @@
+export interface StatType {
+  [cardProperty: string]: string;
+  cardStats: number;
+}
